refactor: missing product component is in its own file
